Address PR review: widget rate limit + robust party lightbulb
Review findings on the always-live session work (PR #2875):

- security: the repurposed /api/widget/take-control re-assert now does a
  board-affecting write (setCurrentClimbAndPublish: optimistic-lock retry
  loop + BLE re-send) but had no rate limit, unlike /api/widget/navigate.
  Extract the per-session token bucket into a shared widget-rate-limit
  module and apply it to both endpoints (capacity 2, refill 1/1.5s); 429
  on drain. Adds a take-control 429 test.
- realtime/ux: the web party lightbulb was driven solely by the
  event-only `wallConfirmed` flag, so a connected member actively relaying
  the wall could show OFF (own bulb before first confirm; darkened when a
  peer's BLE blip broadcast WallDisconnected; late joiner sees OFF). OR in
  the local BLE connection and the board-presence holder, matching mobile's
  PlayDrawer semantics. Solo behavior unchanged.
